+++ src/Settings.Encryption/EncryptDoNotFollowAttribute.cs
@@ -8,8 +8,8 @@ using System;
 namespace Phoenix.Functionality.Settings.Encryption
 {
 	/// <summary>
-	/// Attribute to automatically apply encryption to properties of settings.
+	/// Attribute to not check the properties of the attributed property. This can be used to stop the automatic mechanism from following a nested property recursively.
 	/// </summary>
 	[AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-	public class EncryptAttribute : Attribute { }
+	public class EncryptDoNotFollowAttribute : Attribute { }
 }